requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
aiohttp==3.9.1
asyncpg==0.29.0 